Extract container list
